services/user/get-user/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
PyJWT
pydantic[email]
